feat(775): générer une archive du jeune

+++ src/application/commands/archiver-jeune.command.handler.ts
@@ -1,9 +1,16 @@
 import { Inject, Injectable } from '@nestjs/common'
-import { ArchivageJeune } from 'src/domain/archivage-jeune'
+import {
+  ArchivageJeune,
+  ArchivageJeunesRepositoryToken
+} from 'src/domain/archivage-jeune'
 import { Evenement, EvenementService } from 'src/domain/evenement'
 import { Mail, MailServiceToken } from 'src/domain/mail'
 import { CommandHandler } from '../../building-blocks/types/command-handler'
-import { emptySuccess, Result } from '../../building-blocks/types/result'
+import {
+  emptySuccess,
+  failure,
+  Result
+} from '../../building-blocks/types/result'
 import {
   Authentification,
   AuthentificationRepositoryToken
@@ -13,10 +20,13 @@ import { Chat, ChatRepositoryToken } from '../../domain/chat'
 
 import { Jeune, JeunesRepositoryToken } from '../../domain/jeune'
 import { ConseillerForJeuneAuthorizer } from '../authorizers/authorize-conseiller-for-jeune'
+import { NonTrouveError } from '../../building-blocks/types/domain-error'
+import { DateService } from '../../utils/date-service'
 
 export interface ArchiverJeuneCommand {
   idJeune: Jeune.Id
   motif: ArchivageJeune.Motif
+  commentaire?: string
 }
 
 @Injectable()
@@ -27,6 +37,8 @@ export class ArchiverJeuneCommandHandler extends CommandHandler<
   constructor(
     @Inject(JeunesRepositoryToken)
     private readonly jeuneRepository: Jeune.Repository,
+    @Inject(ArchivageJeunesRepositoryToken)
+    private readonly archivageJeuneRepository: ArchivageJeune.Repository,
     @Inject(ChatRepositoryToken)
     private readonly chatRepository: Chat.Repository,
     @Inject(AuthentificationRepositoryToken)
@@ -35,7 +47,8 @@ export class ArchiverJeuneCommandHandler extends CommandHandler<
     @Inject(MailServiceToken)
     private readonly mailService: Mail.Service,
     private mailFactory: Mail.Factory,
-    private authorizeConseillerForJeune: ConseillerForJeuneAuthorizer
+    private authorizeConseillerForJeune: ConseillerForJeuneAuthorizer,
+    private dateService: DateService
   ) {
     super('ArchiverJeuneCommandHandler')
   }
@@ -51,19 +64,35 @@ export class ArchiverJeuneCommandHandler extends CommandHandler<
   }
 
   async handle(command: ArchiverJeuneCommand): Promise<Result> {
-    const jeune = (await this.jeuneRepository.get(command.idJeune))!
+    const jeune = await this.jeuneRepository.get(command.idJeune)
 
-    await this.authentificationRepository.deleteJeuneIdp(command.idJeune)
-    await this.jeuneRepository.supprimer(command.idJeune)
-    await this.chatRepository.supprimerChat(command.idJeune)
+    if (!jeune) {
+      return failure(new NonTrouveError('Jeune', command.idJeune))
+    }
 
-    if (jeune.conseiller?.email) {
-      const mail = this.mailFactory.creerMailSuppressionJeune(jeune)
-      await this.mailService.envoyer(mail)
-    } else {
-      this.logger.warn(
-        `Email non envoyé au conseiller : ${JSON.stringify(jeune.conseiller)}`
-      )
+    const archive = await this.archivageJeuneRepository.construire(
+      jeune.id,
+      this.dateService.nowJs(),
+      command.motif,
+      command.commentaire
+    )
+    await this.archivageJeuneRepository.archiver(archive!)
+
+    if (false) {
+      await this.authentificationRepository.deleteJeuneIdp(command.idJeune)
+      await this.jeuneRepository.supprimer(command.idJeune)
+      await this.chatRepository.supprimerChat(command.idJeune)
+
+      if (jeune!.conseiller?.email) {
+        const mail = this.mailFactory.creerMailSuppressionJeune(jeune!)
+        await this.mailService.envoyer(mail)
+      } else {
+        this.logger.warn(
+          `Email non envoyé au conseiller : ${JSON.stringify(
+            jeune!.conseiller
+          )}`
+        )
+      }
     }
 
     return emptySuccess()
